+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.5.0] - 2022-11-29
+### Changed
+- Fixed error on setup 'createIntegrationUser'
+- Dropdown for API version
+- Dynamic version number in XML
+- Process JSON response
+
+## [2.4.4] - 2022-11-21
+### Added
+- Reference field contents is now a setting
+- API v1.4 support
+
+### Changed
+- Return layout code for XML
+- Hardcoded API versions to choose from
+- Response content verifier
+
 ## [2.4.0] - 2020-05-28
 ### Added
 - Use native Magento rest and authentication